move elliptic curve actions into common crypto folder
