+++ src/Lyn.Protocol/Bolt3/LightningKeyDerivation.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using Lyn.Protocol.Bolt3.Types;
-using Lyn.Protocol.Common.Hashing;
 using Lyn.Types.Bitcoin;
 using Lyn.Types.Fundamental;
 using NBitcoin;
