+++ src/Lyn.Protocol/Common/DefaultIoCRegistrations.cs
@@ -14,6 +14,7 @@ using Lyn.Protocol.Bolt7.Messages;
 using Lyn.Protocol.Bolt8;
 using Lyn.Protocol.Bolt9;
 using Lyn.Protocol.Common.Blockchain;
+using Lyn.Protocol.Common.Crypto;
 using Lyn.Protocol.Common.Messages;
 using Lyn.Protocol.Connection;
 using Lyn.Types.Serialization;
